online/requirements.txt
flask
tensorflow==2.13.0
opencv-python
pillow